+++ src/main/java/com/library/repository/LigneCreanceRepository.java
@@ -15,11 +15,11 @@ import java.util.List;
 @Repository
 public interface LigneCreanceRepository extends JpaRepository<LigneCreance, Long> {
 
-	List<LigneCreance> findAllByNumero(int numero);
+	List<LigneCreance> findAllByNumero(long numero);
 
 	@Modifying
 	@Query("delete from LigneCreance where numero = :numero")
-	void deleteByNumero(@Param("numero") int numero);
+	void deleteByNumero(@Param("numero") long numero);
 
 	@Query("select p from LigneCreance p where p.produit.id =:prod")
 	List<LigneCreance> ListLigneCreanceByProduitId(@Param("prod") Long prodId);
